ai-service/requirements.txt
# AI Service Dependencies
# Install with: pip install -r requirements.txt

# Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0

# NLP Processing
spacy==3.7.2

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Document Processing
python-docx==1.1.0
python-magic==0.4.27

# OpenAI Integration
openai==1.6.1

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Validation
pydantic==2.5.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
